anomaly-data-generators/requirements.txt
google-cloud-pubsub
google-cloud-storage
Flask
gunicorn